back/api/serializers.py
from rest_framework import serializers
from .models import CustomUser, Disciplina, ReservaAmbiente

class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ['id', 'username', 'email', 'telefone', 'data_nascimento', 'data_contratacao', 'is_professor']

class DisciplinaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Disciplina
        fields = '__all__'

class ReservaAmbienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = ReservaAmbiente
        fields = '__all__'
